Refactored and split ChunkScraper into multiple files
